:recycle: Refactor: tag search api 이동, Photo css 정리
