+++ src/application/hooks/api/search/index.ts
@@ -1,21 +1,8 @@
 import type { QueryFunctionContext } from "@tanstack/react-query";
-import { useInfiniteQuery, useQuery } from "@tanstack/react-query";
+import { useInfiniteQuery } from "@tanstack/react-query";
 
-import { useDebounce } from "@/application/hooks";
 import { api } from "@/infra/api";
-import type { PaginationResponse, SearchItem, Tag } from "@/types";
-
-export const useSearchResult = (value: string) => {
-  const debouncedValue = useDebounce(value);
-
-  const { data, ...rest } = useQuery<{ tags: Tag[] }>({
-    queryKey: ["search", debouncedValue],
-    queryFn: () => api.search.getRecentSearch(debouncedValue),
-    keepPreviousData: true,
-    enabled: !!debouncedValue,
-  });
-  return { searchResults: data?.tags, ...rest };
-};
+import type { PaginationResponse, SearchItem } from "@/types";
 
 const PAGE_SIZE = 20;
 /**
