+++ src/application/hooks/api/tags/index.ts
@@ -1,5 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 
+import { useDebounce } from "@/application/hooks";
 import { api } from "@/infra/api";
 import type { Tag } from "@/types";
 
@@ -10,4 +11,16 @@ export const usePopularTag = () => {
   });
 
   return { tags: data?.tags, ...rest };
+};
+
+export const useGetTagSearch = (value: string) => {
+  const debouncedValue = useDebounce(value);
+
+  const { data, ...rest } = useQuery<{ tags: Tag[] }>({
+    queryKey: ["search", debouncedValue],
+    queryFn: () => api.tags.getTagSearch(debouncedValue),
+    keepPreviousData: true,
+    enabled: !!debouncedValue,
+  });
+  return { searchResults: data?.tags, ...rest };
 };
